# Use optimised merge for merging endpoint parameters