--- a/src/compojure/api/meta.clj
+++ b/src/compojure/api/meta.clj
@@ -11,7 +11,8 @@
             [compojure.api.coerce :as coerce]
             [compojure.api.help :as help]
             compojure.core
-            compojure.api.compojure-compat))
+            compojure.api.compojure-compat
+            [compojure.api.common :as common]))
 
 (def +compojure-api-request+
   "lexically bound ring-request for handlers."
@@ -570,7 +571,7 @@
   "Merge parameters at runtime to allow usage of runtime-paramers with route-macros."
   [{:keys [responses swagger] :as parameters}]
   (cond-> parameters
-          (seq responses) (assoc :responses (apply merge responses))
+          (seq responses) (assoc :responses (common/merge-vector responses))
           swagger (-> (dissoc :swagger) (rsc/deep-merge swagger))))
 
 (defn- route-args? [arg]
@@ -599,7 +600,7 @@
         static? (not (or dynamic? (route-args? route-arg) (seq lets) (seq letks)))
 
         ;; response coercion middleware, why not just code?
-        middleware (if (seq responses) (conj middleware `[coerce/body-coercer-middleware (merge ~@responses)]) middleware)]
+        middleware (if (seq responses) (conj middleware `[coerce/body-coercer-middleware (common/merge-vector ~responses)]) middleware)]
 
     (if context?