Eastwood gives errors from 1-arity merge:

```clj
src/test_eastwood/core.clj:1:1: suspicious-expression: merge called with 1 args. 
(merge map) always returns map.  Perhaps there are misplaced parentheses?
```

use optimised `merge-vector` instead, actually 10x faster for the default case:

```clj
(criterium.core/quick-bench
  (common/merge-vector [{:a 1}]))
; Evaluation count : 84624984 in 6 samples of 14104164 calls.
; Execution time mean : 5.661312 ns
; Execution time std-deviation : 0.305832 ns
; Execution time lower quantile : 5.152271 ns ( 2.5%)
; Execution time upper quantile : 5.942857 ns (97.5%)
; Overhead used : 1.782284 ns

(criterium.core/quick-bench
  (apply merge [{:a 1}]))
; Evaluation count : 8572680 in 6 samples of 1428780 calls.
; Execution time mean : 68.158156 ns
; Execution time std-deviation : 2.062488 ns
; Execution time lower quantile : 65.875315 ns ( 2.5%)
; Execution time upper quantile : 70.444445 ns (97.5%)
; Overhead used : 1.782284 ns
```